+++ ksa_compliance/patches/_2024_06_05_set_cli_setup_to_manual.py
@@ -1,0 +1,5 @@
+import frappe
+
+
+def execute():
+    frappe.db.sql("UPDATE `tabZATCA Business Settings` SET cli_setup = 'Manual'")